Split request handling code out of webcashd.cc into server.cc, so it can be used by the benchmark and test code too.
